backend/requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Validation & Settings
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Redis & Celery
celery[redis]==5.4.0
redis==5.2.1

# AI Providers
openai==1.58.1
anthropic==0.40.0
httpx==0.28.1

# Solidity Tools
py-solc-x==2.0.3

# PDF Generation
reportlab==4.2.5

# Rate Limiting
slowapi==0.1.9

# Logging
structlog==24.4.0
python-json-logger==3.2.1

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0
